Configure an infrastructure that can pre-generate files to quickly run file server performance tests
